Add new trait and typehint ModelInterface
So people can extend whatever they want.
