+++ src/Integration/Symfony/Serializer/DomainDenormalizer.php
@@ -3,7 +3,7 @@
 namespace Biig\Component\Domain\Integration\Symfony\Serializer;
 
 use Biig\Component\Domain\Event\DomainEventDispatcher;
-use Biig\Component\Domain\Model\DomainModel;
+use Biig\Component\Domain\Model\ModelInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 
@@ -27,6 +27,6 @@ final class DomainDenormalizer implements DenormalizerInterface
      */
     public function supportsDenormalization($data, $type, $format = null)
     {
-        return class_exists($type) && is_subclass_of($type, DomainModel::class);
+        return class_exists($type) && is_subclass_of($type, ModelInterface::class);
     }
 }
