+++ src/Model/ModelInterface.php
@@ -2,6 +2,9 @@
 
 namespace Biig\Component\Domain\Model;
 
+use Biig\Component\Domain\Event\DomainEventDispatcher;
+
 interface ModelInterface
 {
+    public function setDispatcher(DomainEventDispatcher $dispatcher);
 }
